requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
WTForms
email-validator
Werkzeug
Pillow
psycopg2-binary
SQLAlchemy
python-dotenv
gunicorn